Update distributing.rst "Development mode" section
Installing project and dependencies in editable mode using local directories, not vcs, using current instructions does not work. The local directory is found, but then the dependency is overridden and fulfilled by PyPI. Changing the order of the editable mode dependencies (i.e. putting editable mode dependencies which reside in local directories at the top of the requirements file) fixes this problem.