Reword cross-reference to the poetry project
The key design distinction between pipenv and poetry isn't
the overall scope of the projects, but rather the assumptions
we make about how projects are going to be structured.

pipenv focuses mainly on "git push" and "git archive" style
publication, where there's a private distribution pipeline
between the repo and production systems (e.g. this is
really common in SaaS and embedded systems).

poetry, by contrast, focuses on the case where a pyproject.toml
file is already present, at which point having Pipfile as well (as
`pipenv` requires if you want to use it to manage development
and testing environments) may feel redundant.