SIngle-sourcing the package version: Remove a py2-only option
Thanks for this guide! I was using this and tried `execfile()` without realizing it didn't work in Python 3.

I thought about leaving the Python 2 version here with a disclaimer, though this list is already very long, and it seems more helpful to steer people toward a good solution than to exhaustively document an obsolete solution.