Remove period punctuation from bullets
Remove period punctuation mark from bullets to make it consistent with
other bullets. Out of four bullets in same section, we're using period
punctuation for only three bullets; and for other section on same page,
we aren't using period punctuation with any bullets, at all. So it would
be nice, if we follow one consistent style through out the
documentations.