Add warning on support of attr: directive
This patch is part of addressing setuptools issue 2492:
https://github.com/pypa/setuptools/issues/2492

The code in this Note block was found to not apply to equivalently-named
features in setup.py.  After discussion on the above issue, a project
maintainer noted this behavior was expected.  This patch adds
documentation of expected behavior close to where it would be discovered
by somebody who stumbled upon the attr: feature, but might not know the
background on directives and setup.cfg vs. setup.py.