fix: Remove '<4' from python_requires in packaging guide
As discussed in https://discuss.python.org/t/use-of-less-than-next-major-version-e-g-4-in-python-requires-setup-py/1066
and other places there can be some side effects of adding `<4` to a
library's python_requires. This PR removes it, as requested in the
discuss.python question.

The section above the lines discussed in the topic that uses compatible
release syntax is also removed given that `~=3.3` also enforces `<4`.

Additional references of problems known to Matthew for posterity:

* Twitter threads:
  - https://twitter.com/HEPfeickert/status/1362170933658738701
  - https://twitter.com/nicholdav/status/1366978826790666240
* GitHub Issues:
  - https://github.com/python-poetry/poetry/issues/3747