Update the "Application dependency management" recommendations
This is a more detailed recommendation that breaks up the recommendation
into multiple parts, to reflect that multiple tools need to be used in
workflows today to maximise compatibility with existing platforms.

It also recommends that deployed applications be isolated from the
system packages, as was implied but not explicitly noted.

The move away from Pipenv as the primary recommendation is a reflection
of the change that application development workflows have a greater
diversity in mature tooling available for them.

Co-authored-by: Brian Rutledge <bhrutledge@gmail.com>