Divorce the section on pip installing extras from setuptools
Per more recent PEPs (e.g. 621), extras may be produced by build tools
other than setuptools if the packager isn't using setuptools as the
build system. In this document, which is mostly about installing,
it seems simpler then to avoid mentioning that setuptools defines
the extras and to simply show the user how to do the install without
mentioning what tool is involved in defining them.