Update pex entry.
Pex long ago stopped offering a programmatic API in favor of its CLI
API. I was remiss in not updating this description. Note is made of the
more expansive features Pex and PEX files have grown in the meantime,
including a nod to Pex multi-platform lock support (RIP PEP 665 - you
live on in Pex).