Clarify that the scripts directory is for files only
Following https://github.com/astral-sh/uv/issues/9656 and the discussion in the pypa discord (https://discord.com/channels/803025117553754132/837243676814999553/1314391181656064040), add a clarifying sentence that the scripts directory in wheels if for files only.

A future specification may want to extend this to "regular files and symlinks", for not it's limited to regular for clear rewrite semantics.