specifications: clarify "sorted, set" to "sorted, collection"
This will hopefully clarify/eliminate a contradiction in terms
in the current specification text: the current text says
both "sorted" and "set", where "set" conventionally implies
an unordered collection.

In context that *appears* to have not been the intention,
since the context is the wheel filename encoding
and not the internal (i.e. parsed) representation
of a wheel filename's tags.

Signed-off-by: William Woodruff <william@yossarian.net>